refactor: move stage veto back to stagelist
