+++ smash/stagelist.py
@@ -1,3 +1,4 @@
+from string import capwords
 from typing import List
 
 from smash.stage import Stage
@@ -21,3 +22,30 @@ class StageList:
                 self.counters.append(Stage(stages[x]['name'],
                                            stages[x]['starter',
                                                      stages[x]['aliases']]))
+
+    def veto(self, stage: str) -> bool:
+        """
+        Tries to veto a stage
+        :param stage: stage to veto
+        :return: boolean value if stage was veto'd
+        """
+        # search for stage in starters
+        for x in range(len(self.starters)):
+            # check for name or alias match
+            if capwords(stage) == self.starters[x].name \
+                    or stage.lower in self.starters[x].aliases:
+                # if matching, change the state of the stage
+                self.starters[x].veto = True
+                return True
+
+        # search through counters
+        for x in range(len(self.counters)):
+            # check for name or alias match
+            if capwords(stage) == self.counters[x].name \
+                    or stage.lower in self.starters[x].aliases:
+                # if matching, change the state of the stage
+                self.counters[x].veto = True
+                return True
+
+        # returns false if it wasn't successful
+        return False
